Remove the moderation namespace
