--- app/controllers/moderation/moderation_controller.rb
@@ -1,11 +1,0 @@
-module Moderation
-  # Base controller all app controllers should inherit from to ensure they inherit the correct settings
-  class ModerationController < ApplicationController
-    protect_from_forgery with: :exception
-    before_action :forbid_non_instance_admins
-
-    def forbid_non_instance_admins
-      render :forbidden, status: :forbidden unless current_user.instance_admin?
-    end
-  end
-end
