--- app/controllers/moderation/pending_projects_controller.rb
@@ -1,7 +1,0 @@
-module Moderation
-  # Web page to approve/reject projects submitted by creators
-  class PendingProjectsController < ModerationController
-    expose(:pending_projects, -> { Project.pending })
-    def index; end
-  end
-end
