--- app/controllers/moderation/project_status_changes_controller.rb
@@ -1,7 +1,0 @@
-module Moderation
-  # Audit log of all the project status changes
-  class ProjectStatusChangesController < ModerationController
-    expose(:project_status_changes, -> { ProjectStatusChange.all.order(created_at: :desc) })
-    def index; end
-  end
-end
