+++ app/controllers/moderation_controller.rb
@@ -1,0 +1,8 @@
+# Parent controller for controllers only accessible to moderators
+class ModerationController < ApplicationController
+  before_action :forbid_non_instance_admins
+
+  def forbid_non_instance_admins
+    render :forbidden, status: :forbidden unless current_user.instance_admin?
+  end
+end
