+++ app/controllers/pending_projects_controller.rb
@@ -1,0 +1,5 @@
+# Interface for admins to review projects for approval/rejection
+class PendingProjectsController < ModerationController
+  expose(:pending_projects, -> { Project.pending })
+  def index; end
+end
