+++ app/controllers/project_status_changes_controller.rb
@@ -1,0 +1,5 @@
+# Audit log of all the project status changes
+class ProjectStatusChangesController < ModerationController
+  expose(:project_status_changes, -> { ProjectStatusChange.all.order(created_at: :desc) })
+  def index; end
+end
