+++ app/views/layouts/application.html.erb
@@ -19,9 +19,8 @@
         <%= link_to t('shared.links.browse_projects'), root_path %>
         <% if current_user %>
           <% if current_user.instance_admin? %>
-            <%= link_to t('moderation.shared.links.pending_projects'), moderation_pending_projects_path %>
-            <%= link_to t('moderation.shared.links.project_status_changes'),
-              moderation_project_status_changes_path %>
+            <%= link_to t('shared.links.pending_projects'), pending_projects_path %>
+            <%= link_to t('shared.links.project_status_changes'), project_status_changes_path %>
           <% end %>
           <%= link_to t('devise.shared.links.sign_out'), destroy_user_session_path, method: :delete %>
         <% else %>
