--- app/views/moderation/project_approvals/new.html.erb
@@ -1,9 +1,0 @@
-<h3><%= pending_project.title %></h3>
-<p><%= pending_project.summary %></p>
-<%= form_for(project_approval,
-             url: moderation_pending_project_project_approvals_url(pending_project),
-             html: { data: { type: "project-status-change-form" } }) do |form| %>
-  <%= form.label :reason, t('moderation.approving_project.reason_field') %><%= form.text_area :reason %>
-  <%= link_to t('moderation.approving_project.cancel'), moderation_pending_projects_path %>
-  <%= form.submit(t('moderation.approving_project.complete')) %>
-<% end %>
