+++ app/views/project_approvals/new.html.erb
@@ -1,0 +1,9 @@
+<h3><%= pending_project.title %></h3>
+<p><%= pending_project.summary %></p>
+<%= form_for(project_approval,
+             url: pending_project_project_approvals_url(pending_project),
+             html: { data: { type: "project-status-change-form" } }) do |form| %>
+  <%= form.label :reason, t('approving_project.reason_field') %><%= form.text_area :reason %>
+  <%= link_to t('approving_project.cancel'), pending_projects_path %>
+  <%= form.submit(t('approving_project.complete')) %>
+<% end %>
