--- features/support/pages/moderation.rb
@@ -1,6 +1,0 @@
-require_relative "moderation/project_status_change_form_section"
-require_relative "moderation/project_status_change_section"
-require_relative "moderation/pending_project_section"
-
-require_relative "moderation/pending_projects_page"
-require_relative "moderation/project_status_changes_page"
