--- features/support/pages/moderation/pending_project_section.rb
@@ -1,6 +1,0 @@
-module Moderation
-  # A single row in the pending project table
-  class PendingProjectSection < SitePrism::Section
-    element :begin_approval_button, "*[data-action=begin-approval]"
-  end
-end
