+++ features/support/pages/pending_project_section.rb
@@ -1,0 +1,4 @@
+# A single row in the pending project table
+class PendingProjectSection < SitePrism::Section
+  element :begin_approval_button, "*[data-action=begin-approval]"
+end
